+++ main.c
@@ -1,0 +1,11 @@
+#include "shell.h"
+/**
+ * main - A simple shell with 0 leaks
+ * Return: 0 when succes
+ */
+
+int main(void)
+{
+	void shell_init();
+	return(0);
+}
