+++ shell.h
@@ -1,5 +1,5 @@
-#ifndef MAIN_H
-#define MAIN_H
+#ifndef SHELL_H
+#define SHELL_H
 
 #include <stdio.h>
 #include <sys/wait.h>
@@ -7,5 +7,9 @@
 #include <unistd.h>
 #include <string.h>
 #include <sys/types.h>
+#include <fcntl.h>
+#include <sys/stat.h>
+
+void shell_init(void);
 
 #endif
